Updated my branch with last master brand updates
