+++ src/main/java/school/faang/user_service/filter/RecommendationRequestFilter/EndDateFilter.java
@@ -1,22 +1,24 @@
-package school.faang.user_service.filter;
+package school.faang.user_service.filter.RecommendationRequestFilter;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
+import school.faang.user_service.filter.Filter;
 
+import java.time.LocalDateTime;
 import java.util.stream.Stream;
 
 @Component
-public class RequesterIdFilter implements Filter<RecommendationRequest, RequestFilterDto> {
+public class EndDateFilter implements Filter<RecommendationRequest, RequestFilterDto> {
 
     @Override
     public boolean isApplicable(RequestFilterDto filter) {
-        return filter.getRequesterId() != null;
+        return filter != null && filter.getEndDate() != null;
     }
 
     @Override
     public Stream<RecommendationRequest> apply(Stream<RecommendationRequest> stream, RequestFilterDto filter) {
-        Long requesterId = filter.getRequesterId();
-        return stream.filter(request -> request.getRequester().getId().equals(requesterId));
+        LocalDateTime endDate = filter.getEndDate();
+        return stream.filter(request -> !request.getCreatedAt().isAfter(endDate));
     }
 }
