+++ src/main/java/school/faang/user_service/filter/RecommendationRequestFilter/RecommendationRequesterIdFilter.java
@@ -1,23 +1,23 @@
-package school.faang.user_service.filter;
+package school.faang.user_service.filter.RecommendationRequestFilter;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
+import school.faang.user_service.filter.Filter;
 
-import java.time.LocalDateTime;
 import java.util.stream.Stream;
 
 @Component
-public class EndDateFilter implements Filter<RecommendationRequest, RequestFilterDto> {
+public class RecommendationRequesterIdFilter implements Filter<RecommendationRequest, RequestFilterDto> {
 
     @Override
     public boolean isApplicable(RequestFilterDto filter) {
-        return filter != null && filter.getEndDate() != null;
+        return filter.getRequesterId() != null;
     }
 
     @Override
     public Stream<RecommendationRequest> apply(Stream<RecommendationRequest> stream, RequestFilterDto filter) {
-        LocalDateTime endDate = filter.getEndDate();
-        return stream.filter(request -> !request.getCreatedAt().isAfter(endDate));
+        Long requesterId = filter.getRequesterId();
+        return stream.filter(request -> request.getRequester().getId().equals(requesterId));
     }
 }
