+++ src/main/java/school/faang/user_service/filter/RecommendationRequestFilter/RecommendationStatusFilter.java
@@ -1,14 +1,15 @@
-package school.faang.user_service.filter;
+package school.faang.user_service.filter.RecommendationRequestFilter;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.RequestStatus;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
+import school.faang.user_service.filter.Filter;
 
 import java.util.stream.Stream;
 
 @Component
-public class StatusFilter implements Filter<RecommendationRequest, RequestFilterDto> {
+public class RecommendationStatusFilter implements Filter<RecommendationRequest, RequestFilterDto> {
 
     @Override
     public boolean isApplicable(RequestFilterDto filter) {
